# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Locadora de Filmes",
    "description": "API para gerenciar dados de uma locadora de filmes, incluindo filmes, atores, gêneros, personagens e classificações.",
    "contact": {
      "name": "João Pedro Teodoro Nunes Correia"
    }
  },
  "host": "localhost:8080",
  "basePath": "/v1/locadora",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/locadora/filme": {
      "get": {
        "tags": [
          "Filme"
        ],
        "summary": "Listar todos os filmes",
        "description": "Endpoint para listar todos os filmes cadastrados.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Filme"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Não encontrado",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Erro interno",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Filme"
        ],
        "summary": "Inserir um novo filme",
        "description": "Endpoint para cadastrar um novo filme.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados do filme",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Criado com sucesso",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        }
      }
    },
    "/v1/locadora/filme/{id}": {
      "get": {
        "tags": [
          "Filme"
        ],
        "summary": "Buscar filme por ID",
        "description": "Endpoint para buscar um filme pelo seu ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          },
          "400": {
            "description": "ID inválido",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "Não encontrado",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Filme"
        ],
        "summary": "Atualizar um filme",
        "description": "Endpoint para atualizar os dados de um filme existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do filme"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados para atualização",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Atualizado com sucesso",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Filme"
        ],
        "summary": "Excluir um filme",
        "description": "Endpoint para excluir um filme pelo seu ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Excluído com sucesso"
          }
        }
      }
    },
    "/v1/locadora/personagem": {
      "get": {
        "tags": [
          "Personagem"
        ],
        "summary": "Listar todos os personagens",
        "description": "Endpoint para listar todos os personagens.",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Personagem"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Personagem"
        ],
        "summary": "Inserir um novo personagem",
        "description": "Endpoint para cadastrar um novo personagem.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados do personagem",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Personagem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/locadora/personagem/{id}": {
      "get": {
        "tags": [
          "Personagem"
        ],
        "summary": "Buscar personagem por ID",
        "description": "Endpoint para buscar um personagem pelo seu ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do personagem"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Personagem"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Personagem"
        ],
        "summary": "Atualizar um personagem",
        "description": "Endpoint para atualizar os dados de um personagem.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do personagem"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados para atualização",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Personagem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Personagem"
        ],
        "summary": "Excluir um personagem",
        "description": "Endpoint para excluir um personagem pelo seu ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do personagem"
          }
        ],
        "responses": {
          "200": {
            "description": "Excluído com sucesso"
          }
        }
      }
    },
    "/v1/locadora/genero": {
      "get": {
        "tags": [
          "Gênero"
        ],
        "summary": "Listar todos os gêneros",
        "description": "Endpoint para listar todos os gêneros.",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Genero"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Gênero"
        ],
        "summary": "Inserir um novo gênero",
        "description": "Endpoint para cadastrar um novo gênero.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados do gênero",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Genero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/locadora/genero/filme/{id}": {
      "get": {
        "tags": [
          "Gênero"
        ],
        "summary": "Listar gêneros de um filme específico",
        "description": "Endpoint para listar todos os gêneros associados a um filme pelo ID do filme.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Genero"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/v1/locadora/classificacao": {
      "get": {
        "tags": [
          "Classificação"
        ],
        "summary": "Listar todas as classificações",
        "description": "Endpoint para listar todas as classificações indicativas.",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Classificacao"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Classificação"
        ],
        "summary": "Inserir uma nova classificação",
        "description": "Endpoint para cadastrar uma nova classificação.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados da classificação",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Classificacao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "boolean"
            },
            "example": {
              "type": "boolean",
              "example": false
            }
          }
        },
        "status_code": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "integer"
            },
            "example": {
              "type": "number",
              "example": 404
            }
          }
        },
        "message": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "string"
            },
            "example": {
              "type": "string",
              "example": "Não foram encontrados dados de retorno!"
            }
          }
        }
      }
    }
  }
}